feat(mps-sync-plugin): status bar widget
